♻ refactor: controller 이름 수정 & dto 파일 구조 변경 & 코드 정리
