+++ src/main/java/org/scoula/domain/precontract/dto/tenant/TenantStep3DTO.java
@@ -1,4 +1,4 @@
-package org.scoula.domain.precontract.dto;
+package org.scoula.domain.precontract.dto.tenant;
 
 import org.scoula.global.common.constant.Constants;
